# public/swagger.yaml
openapi: 3.0.3
info:
  title: Food Delivery System API
  description: |
    RESTful API for Food Delivery System with RAG Chatbot

    ## Features
    - User authentication & authorization
    - Restaurant & menu management
    - Order processing & tracking
    - Real-time delivery tracking
    - Push notifications
    - Chat messaging
    - RAG-powered chatbot

    ## Base URL
    - Development: `http://localhost:3000`
  version: 1.0.0
  contact:
    name: API Support
    email: support@fooddelivery.com
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/orders/{orderId}:
    patch:
      tags:
        - Orders
      summary: Update order status
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [PENDING, ACCEPTED, PREPARING, READY, PICKED_UP, ON_THE_WAY, DELIVERED, CANCELLED, REJECTED]
      responses:
        '200':
          description: Order updated
...
  /api/notifications:
    get:
      tags:
        - Notifications
      summary: Get user notifications
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of notifications
...